chore: generalize PaMessageForm
Moves form component for PA messages from `NewPaMessage` into
`PaMessageForm` so that it can be used to make an `EditPaMessage`
component later on.
