+++ assets/js/components/Dashboard/PaMessageForm/SelectStationsAndZones/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SelectStationsAndZones";
