+++ assets/js/components/Dashboard/PaMessageForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PaMessageForm";
