+++ assets/js/utils/api.ts
@@ -4,7 +4,7 @@ import { ScreenConfiguration } from "../models/screen_configuration";
 import { ScreensByAlert } from "../models/screensByAlert";
 import { PlaceIdsAndNewScreens } from "../components/Dashboard/PermanentConfiguration/Workflows/GlEink/ConfigureScreensPage";
 import getCsrfToken from "../csrf";
-import { PaMessage } from "Models/pa_message";
+import { NewPaMessageBody } from "Models/pa_message";
 
 export const fetchPlaces = async (): Promise<Place[]> => {
   const response = await fetch("/api/dashboard");
@@ -117,7 +117,7 @@ export const publishScreensForPlace = async (
   return { status: response.status, message };
 };
 
-export const createNewPaMessage = async (message: PaMessage) => {
+export const createNewPaMessage = async (message: NewPaMessageBody) => {
   const response = await fetch("/api/pa-messages", {
     ...getPostBodyAndHeaders(message),
     credentials: "include",
